+++ presentation/src/main/kotlin/me/proton/core/presentation/ui/ProtonSecureActivity.kt
@@ -18,10 +18,10 @@
 
 package me.proton.core.presentation.ui
 
-import android.os.Bundle
 import android.view.LayoutInflater
-import android.view.WindowManager
 import androidx.viewbinding.ViewBinding
+import me.proton.core.presentation.utils.ProtectScreenConfiguration
+import me.proton.core.presentation.utils.protectScreen
 
 abstract class ProtonSecureActivity<ViewBindingT: ViewBinding>(
     bindingInflater: (LayoutInflater) -> ViewBindingT
@@ -29,26 +29,8 @@ abstract class ProtonSecureActivity<ViewBindingT: ViewBinding>(
 
     /** Can be overridden to modify the protections applied to the extending activity */
     @Suppress("MemberVisibilityCanBePrivate")
-    protected val configuration = Configuration(preventScreenRecording = true)
+    protected val configuration = ProtectScreenConfiguration()
 
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
+    private val screenProtector by protectScreen(configuration)
 
-        protectActivity()
-    }
-
-    /** Sets up the necessary protections based on [configuration] */
-    private fun protectActivity() {
-        if (configuration.preventScreenRecording) preventScreenRecording()
-    }
-
-    private fun preventScreenRecording() {
-        window.addFlags(WindowManager.LayoutParams.FLAG_SECURE)
-    }
-
-    /** Used to configure which protections should be applied to the activity */
-    class Configuration(
-        /** Prevent screen capture etc. to record user password. Also masks app in app switcher. */
-        val preventScreenRecording: Boolean
-    )
 }